Move bundle & sign related ELF operations to elf.go
